Update README.md
# Jackson_Final_Project
This purpose of this project is to compare patient chronic condition groupings from a Medicaid health plan which is interested in population care management and Triple Aim: improving population health, improving patient care, and reducting the per capita cost of provideing health care.  These goals can be achieved by looking at patient chronic condition groupings that have the following attributes:  the name of the chronic condtion, the number of members, overall cost per member, drug and medical cost per member, admissions, ED visits, and total cost.  Determining the cost of treating particular chronic condition populations and what drives the cost (the attributes) can help a health plan achieve the Triple Aim.  This project uses chronic conditions, age groups, ED visits, total cost as the starting point for comparing chronic conditon costs.  The visual shows chronic condtion groupings whose total costs are greater than $200,000.00.

The project can be used to compare chronic conditions by total cost, drug cost, medical cost, ED visits, admissions, EBM complicance, number of members.  The project can also be used to compare by age group:  50-64, 65-84, over 85.


The modules that need to be imported are:  csv, pandas, matplotlib.pyplot as plt, numpy as np.


MIT License

Copyright (c) [2017] [Natalie Jackson]

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.